Split xformers.ops
[ghstack-poisoned]
